+++ src/common/ipc-keys.ts
@@ -1,0 +1,11 @@
+/**
+ * IPC keys.
+ */
+
+ export enum IPCWindowKey {
+  SendMessage = 'SendMessage',
+  CreateNewWindow = 'CreateNewWindow',
+  GetWindowIds = 'GetWindowIds',
+  UpdateWindowIds = 'UpdateWindowIds',
+  UpdateMessage = 'UpdateMessage'
+}