+++ src/renderer/views/primary/app/core/core.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+
+/**
+ * App Core module
+ * ! Important: import only once in AppModule
+ * ! All services should be singletons
+ * @see https://angular.io/guide/singleton-services
+ */
+@NgModule({
+  declarations: [],
+  imports: [
+    CommonModule
+  ]
+})
+export class CoreModule { }