+++ src/renderer/views/primary/app/shared/shared.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { TranslateModule } from '@ngx-translate/core';
+
+import { PageNotFoundComponent } from './components/';
+import { WebviewDirective } from './directives/';
+import { FormsModule } from '@angular/forms';
+
+/**
+ * Shared module.
+ * Import this module in other modules if you need to
+ * share the same component / directive between multiple modules.
+ */
+@NgModule({
+  declarations: [PageNotFoundComponent, WebviewDirective],
+  imports: [CommonModule, TranslateModule, FormsModule],
+  exports: [TranslateModule, WebviewDirective, FormsModule]
+})
+export class SharedModule { }