+++ src/renderer/views/primary/main.ts
@@ -1,0 +1,21 @@
+import { enableProdMode } from '@angular/core';
+import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+
+import { AppModule } from './app/app.module';
+import { AppConfig } from './environments/environment';
+
+/**
+ * Enable production mode
+ */
+if (AppConfig.production) {
+  enableProdMode();
+}
+
+/**
+ * Bootstrap App module
+ */
+platformBrowserDynamic()
+  .bootstrapModule(AppModule, {
+    preserveWhitespaces: false
+  })
+  .catch(err => console.error(err));